generate case control annotations for hit data based on library annotation file.

generate case control annotations for hit data based on library annotation file.

AnnotationGenerator(data, data_level, data_types, annot,
  peptide_col_id_match, protein_col_id_match, peptide_col_id_display,
  protein_col_id_display, description_col_id, flag_col_id, flag_type,
  pep_aa = "pep_aa")

Arguments

data
Input data frame.

data_level
If data_level is "peptide", then AnnotationGenerator will operate based on the peptide_ parameters. If data_level is "protein", then Annotation_Generator will operate based on the protein_ parameters.

data_types
Character vector describing types of data stored in data list. e.g. "zscore", "polyclonal".

annot
Library annotation file with named columns.

peptide_col_id_match
For data_level=="peptide", the name of a column in annot that corresponds to the first column of data.

protein_col_id_match
For data_level=="protein", the name of a column in annot that corresponds to the first column of data.

protein_col_id_display
The name of a column in annot that corresponds to identifiers to display in the markdown report corresponding to each protein.

description_col_id
The name of a column in annot that corresponds to identifiers to display in the markdown report corresponding to gene product or similar short protein descriptions to display in the markdown report.

flag_col_id
Character string or vector of column name(s) corresponding to annotation flags to indicate in the markdown report (e.g. of known autoantigens or surface proteins).

flag_type
type "binary" to detect whether or not there is text in that cell. "literal" to display exact text from cell.

pep_aa
Character string of column name in annot corresponding to peptide amino acid sequence in the markdown report. Needed for peitopefindr.

peptide_cold_id_display
For data_level=="peptide", the name of a column in annot that corresponds to identifiers to display in the markdown report corresponding to each peptide.