Pass parameters to R markdown case control template and knit html output file.

Pass parameters to R markdown case control template and knit html output file.

render_from_template(template = system.file("template_case_control.Rmd",
  package = "phipcc"), output_file = NULL, output_dir = getwd(),
  set_title = NULL, import_rdata = FALSE,
  rdata_path = "data/outputs.RData", config = NULL,
  candidate_table_flagged = NULL, candidate_table_full = NULL,
  AVARDA_candidate_table = NULL, graphs = NULL, clustergram1 = NULL,
  motifs = NULL, clustergram2 = NULL)

Arguments

template
Rmd output template file.

output_file
Name of generated html report. If not provided, a default value is generated.

output_dir
Directory in which to write report.

set_title
Character string to be used for report title. If not provided, a default value is generated.

import_rdata
Logical whether or not to look for an .rdata file containing necessary targests at the path specified in rdata_path.

rdata_path
If import_rdata == TRUE, then an file with the specified name will be passed to load().

config
Configuration table used in phipcc:define_plan_case_control. After running drake::make on this plan, config can be retreived by calling drake::readd(config). If not provided, a default value is generated.

candidate_table_flagged
Kable html object of flagged/annotated candidate hits to display in report. Retrieved by calling drake::readd(candidate_table_flagged_html). If not provided, a default value is generated.

candidate_table_full
Kable html object of all candidate hits to display in report. Retrieved by calling drake::readd(candidate_table_html). If not provided, a default value is generated.

AVARDA_candidate_table
Kable html object of all AVARDA hits to display in report. Retrieved by calling drake::readd(AVARDA_candidate_table). If not provided, a default value is generated.

graphs
List of ggplot objects to be plotted in report. Retreived by calling list(drake::readd(plot1), drake::readd(plot2), drake::readd(plot3)). If not provided, a default value is generated.

clustergram1
List of gplots::heatmap.2 object and sorted data. Retreived by calling drake::readd(clustergram1). If not provided, a default value is generated.

motifs
Directory path to epitopefindr motifs folder. Typically paste0(getwd(), "/data/epitopefindr/intermediate_files/msa/"). If not provided, a default value is generated.

clustergram2
List of gplots::heatma.2 object after collapsing epitopefindr groups. Retreived by calling drake::readd(clustergram2). If not provided, a default value is generated.

env
Environment from which to load drake objects.

See also